Add support for hiding character groups on Collide and Act 7
Collide and Act 7 seemingly can't be handled in the same way as every
other page, so the style was hardcoded in collideact7.css.  As a
result, changing the settings (hiding a character group, or always
displaying character names) would have no effect on those pages.

With this commit, the style for Collide and Act 7 is dynamically
calculated, and then applied on the next start.  As a result, changing
the settings requires restarting the app twice for the changes to be
visible.  Which is approximately infinitely less times than before.

Still, it's not ideal.