Use characterNextLinks for index, nth-last-child, CSS cont. [2/?]
Instead of basing the index on the length of the next pages array for
a page, we instead calculate ahead of time the next pages and use the
length of that as the index counting from the bottom.

This change is made to support the Terezi retcons, which, if hidden,
removed the first link from the page and made the previously
calculated nth-child value off by one.  As a result, the links were
labeled incorrectly.  By counting from the bottom, we don't need to
care that the first link may be removed.