Don't add links when character is hidden [3/3]
Previously, links for hidden characters were added to the next pages
array, then hidden with CSS.  This worked with the previously because
the index was still correct: no matter what characters were hidden,
the index for any character remained the same.

With the new system where the last character's CSS is added first, the
index would no longer match if a character was hidden with; they had
still been added as divs to the page.

With this change, rather than adding a link and then hiding it, the
link just won't be added to begin with.

As a bonus, this also fixes arrow key navigation when there are only
hidden characters on a page.  Previously, there would be multiple
links on the page (they'd just be hidden), so the arrow key wouldn't
know which link to go to.  Now that the links aren't added, the arrow
key goes to the only link on the page.