Allow overriding a link's name, colour, and image
Unlike the ordinanry "Name", "Colour", and "Image", the override
variations "Name Override", "Colour Override", and "Image Override"
style the link on the current page, and that link alone.  By contrast,
the ordinary variation will not style the link, but will just set the
name or colour or image for the next page onward.  For example:

  Name: John
  1901-1902

  Name Override: Zoosmell Pooplord
  1903

will change the appearance of the link on page 1902 to have the name
"Zoosmell Pooplord" appear when one hovers over the link instead of
"John" like normal, and the link for John on page 1903 will say "John"
like normal.  By contrast, with this:

  Name: John
  1901-1902

  Name: Zoosmell Pooplord
  1903

the name that appears upon hovering over the link on page 1902 will
still say "John", and the links for John on page 1903 and onward will
say "Zoosmell Pooplord".

Replaces Caption, which was used only three times.